Move _import_dir_helper and _export_dir_helper from workspace.cli to workspace.api (#130)
* changes

* Moved _export_dir_helper and _import_dir_helper from workspace/cli.py to workspace/api.py and renamed to import_workspace_dir and export_workspace_dir. added import_workpsace_dir and export_workspace_dir tests in api tests, need to address lint

* Fixed lint
